Create ListBooks component
ListBooks is a functional component which expects to be provided with
`books` (list of books to be rendered), `shelves` (map of the shelves
with book-ids) and `onShelveUpdate` (callback function for moving the
book between the shleves). ListBooks internally renders another
functional component: `ShelfChanger`, which is responsible for rendering
the select menu with bookshelves.