Create SearchBooks component
This component is responsible for providing a functionality for
searching a book or set of books based on a given query. It is expected
to be called with the `shelves` (BooksApp.categories) and
`onShelfUpdate` as it should be possible to change the shelf of the book
on the search page too.