Add initial setup to the BooksApp
Once BooksApp component was mounted, initial setup needs to be
performed. It consists of calling the BooksAPI.getAll() and changing the
state of the app. All of the books which have the shelf selected are
fetched from the database and the state of the app is then changed based
on the result of the api call.